Implement WASM Support with increasing okhttp versions (#913)
* adding the wasm support

* Rename HttpLoader.wasmJs.kt to HttpLoader.kt

* Rename HttpLoader.kt to WasmHttpLoader.kt

* Update HttpUrlFetcher.kt

* Update Main.kt

* merge latest changes from main

* Update Object.kt

* Update HttpUrlFetcher.kt

* Update HttpUrlFetcher.kt

+++ fuel/src/wasmJsMain/kotlin/fuel/HttpResponse.wasmJs.kt
@@ -3,7 +3,7 @@ package fuel
 import org.w3c.fetch.Response
 
 public actual class HttpResponse {
-    public var statusCode: Int = -1
+    public var statusCode: Short = -1
     public var response: Response? = null
     public var headers: Map<String, String> = emptyMap()
 }
